Update python_sdk_examples.ipynb
The notebook file had a JSON compilation error, making it impossible to open and view on a system. It turned out to be a missing comma, which could only be identified by running the .ipynb file through a JSON validator and compiling it to find the error. This is a tedious task, and I believe correcting it would save time for many people.

Thank you.